Renamed browser to app; moved css to static files; fixed click args not working in hot-reloads
