+++ isoplots/isonice/static/styles.css
@@ -1,0 +1,11 @@
+.hide-header .ag-header {
+    display: none;
+}
+
+.nicegui-content {
+    padding: 0;
+}
+
+.flex-grow .q-field__control {
+    height: 100%;
+}
